New benchmarking harness with sum benchmark [note]
The harness can be used to execute a module for benchmarking. The harness first loads the module via the SOM platform and then extracts all methods, of the loaded module, whose signatures begin with "benchmark". Each method is then executed and the results of the benchmarked logged. The execution time can be present in either a human-readable form (use the argument `pretty`) or as CSV (use the argument `csv`).