Add Harness [note]
The harness is used to time the execution of a benchmark module. The harness first loads the module and then invokes and times it's `benchmark(_)` methods. The execution time is presented in a TSV format. When invoked without additional arguments the harness runs each of the seven micro benchmarks introduced here. A particular benchmark may be specified by adding additional arguments.